+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/InitDBConfiguration.java
@@ -29,8 +29,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
-import uk.ac.ebi.eva.pipeline.steps.GenesLoad;
-import uk.ac.ebi.eva.pipeline.steps.tasklet.IndexesCreate;
+import uk.ac.ebi.eva.pipeline.jobs.steps.GenesLoad;
+import uk.ac.ebi.eva.pipeline.jobs.steps.tasklet.IndexesCreate;
 
 @Configuration
 @EnableBatchProcessing
