+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/VariantsAnnotGenerateInput.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.pipeline.steps;
+package uk.ac.ebi.eva.pipeline.jobs.steps;
 
 import com.mongodb.DBObject;
