+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/tasklet/VariantsLoad.java
@@ -13,15 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.pipeline.steps.tasklet;
+package uk.ac.ebi.eva.pipeline.jobs.steps.tasklet;
 
-import org.opencb.biodata.models.variant.VariantSource;
 import org.opencb.datastore.core.ObjectMap;
-import org.opencb.datastore.core.QueryOptions;
 import org.opencb.opencga.storage.core.StorageManagerFactory;
 import org.opencb.opencga.storage.core.variant.VariantStorageManager;
-import org.opencb.opencga.storage.core.variant.adaptors.VariantDBAdaptor;
-import org.opencb.opencga.storage.core.variant.stats.VariantStatisticsManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.batch.core.StepContribution;
@@ -37,6 +33,8 @@ import uk.ac.ebi.eva.VariantJobsArgs;
 import uk.ac.ebi.eva.utils.URLHelper;
 
 import java.net.URI;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 /**
  * Created by jmmut on 2015-11-10.
@@ -46,8 +44,8 @@ import java.net.URI;
 @Component
 @StepScope
 @Import({VariantJobsArgs.class})
-public class VariantsStatsLoad implements Tasklet {
-    private static final Logger logger = LoggerFactory.getLogger(VariantsStatsLoad.class);
+public class VariantsLoad implements Tasklet {
+    private static final Logger logger = LoggerFactory.getLogger(VariantsLoad.class);
 
     @Autowired
     private VariantJobsArgs variantJobsArgs;
@@ -57,17 +55,25 @@ public class VariantsStatsLoad implements Tasklet {
         ObjectMap variantOptions = variantJobsArgs.getVariantOptions();
         ObjectMap pipelineOptions = variantJobsArgs.getPipelineOptions();
 
-        VariantStorageManager variantStorageManager = StorageManagerFactory.getVariantStorageManager();
-        QueryOptions statsOptions = new QueryOptions(variantOptions);
-        VariantStatisticsManager variantStatisticsManager = new VariantStatisticsManager();
-        VariantDBAdaptor dbAdaptor = variantStorageManager.getDBAdaptor(variantOptions.getString("dbName"), variantOptions);
+        VariantStorageManager variantStorageManager = StorageManagerFactory.getVariantStorageManager();// TODO add mongo
         URI outdirUri = URLHelper.createUri(pipelineOptions.getString("output.dir"));
-        VariantSource variantSource = variantOptions.get(VariantStorageManager.VARIANT_SOURCE, VariantSource.class);
-        URI statsOutputUri = outdirUri.resolve(VariantStorageManager.buildFilename(variantSource));
+        URI nextFileUri = URLHelper.createUri(pipelineOptions.getString("input.vcf"));
 
-        // actual stats load
-        variantStatisticsManager.loadStats(dbAdaptor, statsOutputUri, statsOptions);
+//          URI pedigreeUri = pipelineOptions.getString("input.pedigree") != null ? createUri(pipelineOptions.getString("input.pedigree")) : null;
+        Path output = Paths.get(outdirUri.getPath());
+        Path input = Paths.get(nextFileUri.getPath());
+        Path outputVariantJsonFile = output.resolve(input.getFileName().toString() + ".variants.json" + pipelineOptions.getString("compressExtension"));
+//          outputFileJsonFile = output.resolve(input.getFileName().toString() + ".file.json" + config.compressExtension);
+        URI transformedVariantsUri = outdirUri.resolve(outputVariantJsonFile.getFileName().toString());
+
+        logger.info("-- PreLoad variants -- {}", nextFileUri);
+        variantStorageManager.preLoad(transformedVariantsUri, outdirUri, variantOptions);
+        logger.info("-- Load variants -- {}", nextFileUri);
+        variantStorageManager.load(transformedVariantsUri, variantOptions);
+//          logger.info("-- PostLoad variants -- {}", nextFileUri);
+//          variantStorageManager.postLoad(transformedVariantsUri, outdirUri, variantOptions);
 
         return RepeatStatus.FINISHED;
     }
+
 }
