+++ src/test/java/uk/ac/ebi/eva/pipeline/jobs/VariantConfigurationWorkflowTest.java
@@ -18,12 +18,9 @@ package uk.ac.ebi.eva.pipeline.jobs;
 
 import uk.ac.ebi.eva.VariantJobsArgs;
 import uk.ac.ebi.eva.pipeline.config.VariantWorkflowConfig;
-import uk.ac.ebi.eva.pipeline.jobs.VariantAnnotConfiguration;
-import uk.ac.ebi.eva.pipeline.jobs.VariantConfiguration;
-import uk.ac.ebi.eva.pipeline.jobs.VariantStatsConfiguration;
-import uk.ac.ebi.eva.pipeline.steps.VariantsAnnotGenerateInput;
-import uk.ac.ebi.eva.pipeline.steps.VariantsAnnotLoad;
-import uk.ac.ebi.eva.pipeline.steps.tasklet.VariantsAnnotCreate;
+import uk.ac.ebi.eva.pipeline.jobs.steps.VariantsAnnotGenerateInput;
+import uk.ac.ebi.eva.pipeline.jobs.steps.VariantsAnnotLoad;
+import uk.ac.ebi.eva.pipeline.jobs.steps.tasklet.VariantsAnnotCreate;
 
 import java.io.File;
 import java.nio.file.Paths;
